+++ src/main/java/org/delia/db/newhls/HLDField.java
@@ -1,0 +1,27 @@
+package org.delia.db.newhls;
+
+import org.delia.type.BuiltInTypes;
+import org.delia.type.DStructType;
+import org.delia.type.DType;
+
+/**
+	 * Represents a field mentioned in a query. Has all the info for rendering SQL or doing MEM query.
+	 * @author ian
+	 *
+	 */
+	public class HLDField {
+		public DStructType structType;
+		public String fieldName;
+		public DType fieldType;
+		//		public boolean isAssocField;
+//		public String groupName; just use alias??
+		public String alias;
+		public String asStr;
+
+		@Override
+		public String toString() {
+			String fldType = BuiltInTypes.convertDTypeNameToDeliaName(fieldType.getName());
+			String s = String.format("%s.%s(%s,%s)", structType.getName(), fieldName, fldType, alias);
+			return s;
+		}
+	}
