+++ src/main/java/org/delia/db/newhls/QueryFnSpec.java
@@ -1,0 +1,19 @@
+package org.delia.db.newhls;
+
+import org.delia.db.newhls.cond.FilterFunc;
+
+/**
+ * A function such as orderBy,count,first,min,etc
+ * @author ian
+ *
+ */
+public class QueryFnSpec {
+	public RelationField structField; //who the func is being applied to. fieldName & fieldType can be null
+	public FilterFunc filterFn;
+
+	@Override
+	public String toString() {
+		String s = String.format("%s %", structField.toString(), filterFn.toString());
+		return s;
+	}
+}
