+++ src/test/java/org/delia/db/newhls/NewHLSTests.java
@@ -3,9 +3,7 @@ package org.delia.db.newhls;
 
 import static org.junit.Assert.assertEquals;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 import java.util.StringJoiner;
 
 import org.delia.compiler.ast.QueryExp;
@@ -24,10 +22,8 @@ import org.delia.db.newhls.cond.StringFilterCond;
 import org.delia.db.sql.StrCreator;
 import org.delia.relation.RelationCardinality;
 import org.delia.relation.RelationInfo;
-import org.delia.type.BuiltInTypes;
 import org.delia.type.DStructType;
 import org.delia.type.DType;
-import org.delia.type.DTypeRegistry;
 import org.delia.type.TypePair;
 import org.delia.util.DRuleHelper;
 import org.delia.util.DValueHelper;
@@ -116,279 +112,6 @@ public class NewHLSTests extends HLSTestBase {
 	//	 * listFn          orderBy/distinct/limit/offset/first,last,ith      sort,paging. optional    
 	//	 * fetch           fks,fetch('aaa'),...          load 0 or more sub-objects.
 	//	 * calcFn          exists,count,min,max,average,sum,...   query produces a calculated result. 
-
-	public static class FetchSpec {
-		public RelationField structField;
-		public boolean isFK; //if true then fks, else fetch
-
-		@Override
-		public String toString() {
-			String s = String.format("%s:%b", structField.toString(), isFK);
-			return s;
-		}
-	}
-	public static class QueryFnSpec {
-		public RelationField structField; //who the func is being applied to. fieldName & fieldType can be null
-		public FilterFunc filterFn;
-
-		@Override
-		public String toString() {
-			String s = String.format("%s %", structField.toString(), filterFn.toString());
-			return s;
-		}
-	}
-
-	public static class JoinElement  {
-		public RelationField structField;
-		//		public List<JTElement> nextL = new ArrayList<>();
-		public RelationInfo relinfo;
-		public boolean usedForFK; //if true then fks(). but this join for other reasons too
-		public boolean usedForFetch; //if true then fettch. but this join for other reasons too
-
-		@Override
-		public String toString() {
-			StringJoiner joiner = new StringJoiner("|");
-			joiner.add(structField.dtype.getName());
-			joiner.add(structField.fieldName);
-			joiner.add(structField.fieldType.getName());
-			return joiner.toString();
-		}
-
-		public boolean matches(TypePair pair) {
-			if (pair.name.equals(structField.fieldName) && pair.type == structField.fieldType) {
-				return true;
-			}
-			return false;
-		}
-
-		public TypePair createPair() {
-			return new TypePair(structField.fieldName, structField.fieldType);
-		}
-	}	
-	public static class HLDQuery {
-		public DStructType fromType;
-		public String fromAlias;
-		public DStructType mainStructType; //C[].addr then fromType is A and mainStringType is C
-		public DType resultType; //might be string if .firstName
-		public FilterCond filter;
-		public List<RelationField> throughChain = new ArrayList<>();
-		public StructField finalField; //eg Customer.addr
-		public List<FetchSpec> fetchL = new ArrayList<>(); //order matters: eg. .addr.fetch('country')
-		public List<QueryFnSpec> funcL = new ArrayList<>(); //list and calc fns. order matters: eg. .addr.first().city
-		public List<HLDField> fieldL = new ArrayList<>(); 
-
-		//added after
-		public List<JoinElement> joinL = new ArrayList<>();
-
-		@Override
-		public String toString() {
-			String s = String.format("%s:%s[]:%s", resultType.getName(), fromType.getName(), mainStructType.getName());
-			s += String.format(" [%s]", filter.toString());
-
-			if (throughChain.isEmpty()) {
-				s += " TC[]";
-			} else {
-				StringJoiner joiner = new StringJoiner(",");
-				for(RelationField sf: throughChain) {
-					joiner.add(sf.toString());
-				}
-				s += String.format(" TC[%s]", joiner.toString());
-			}
-
-			if (finalField == null) {
-				s += " fld()";
-			} else {
-				s += String.format(" fld(%s)", finalField.toString());
-			}
-
-			if (fetchL.isEmpty()) {
-				s += " fetch[]";
-			} else {
-				StringJoiner joiner = new StringJoiner(",");
-				for(FetchSpec sf: fetchL) {
-					joiner.add(sf.toString());
-				}
-				s += String.format(" fetch[%s]", joiner.toString());
-			}
-
-			if (funcL.isEmpty()) {
-				s += " fn[]";
-			} else {
-				StringJoiner joiner = new StringJoiner(",");
-				for(QueryFnSpec sf: funcL) {
-					joiner.add(sf.toString());
-				}
-				s += String.format(" fn[%s]", joiner.toString());
-			}
-
-			if (fieldL.isEmpty()) {
-				s += " {}";
-			} else {
-				StringJoiner joiner = new StringJoiner(",");
-				for(HLDField rf: fieldL) {
-					joiner.add(rf.toString());
-				}
-				s += String.format(" {%s}", joiner.toString());
-			}
-			return s;
-		}
-	}
-
-	public static class HLDQueryBuilder {
-		private DTypeRegistry registry;
-
-		public HLDQueryBuilder(DTypeRegistry registry) {
-			this.registry = registry;
-		}
-
-		public HLDQuery build(QueryExp queryExp) {
-			HLDQuery hld = new HLDQuery();
-			hld.fromType = (DStructType) registry.getType(queryExp.typeName);
-			hld.mainStructType = hld.fromType; //TODO fix
-			hld.resultType = hld.fromType; //TODO fix
-
-			FilterCondBuilder builder = new FilterCondBuilder();
-			hld.filter = builder.build(queryExp);
-			//			public List<StructField> throughChain = new ArrayList<>();
-			//			public StructField finalField; //eg Customer.addr
-			//			public List<FetchSpec> fetchL = new ArrayList<>(); //order matters: eg. .addr.fetch('country')
-			//			public List<QueryFnSpec> funcL = new ArrayList<>(); //list and calc fns. order matters: eg. .addr.first().city
-
-			return hld;
-		}
-	}
-
-	public static class JoinTreeBuilder {
-		public void generateJoinTree(HLDQuery hld) {
-			for(FetchSpec spec: hld.fetchL) {
-				if (spec.isFK) {
-					addFKs(spec, hld.joinL);
-				} else {
-					addFetch(spec, hld.joinL);
-				}
-			}
-
-			for(QueryFnSpec fnspec: hld.funcL) {
-				//if parent does .orderBy('addr') then we need a join.
-				if (fnspec.filterFn.fnName.equals("orderBy")) {
-					addOrderBy(fnspec, hld.joinL);
-				}
-			}
-
-			//TODO: add throughChain
-
-			if (hld.filter instanceof OpFilterCond) {
-				OpFilterCond ofc = (OpFilterCond) hld.filter;
-				addImplicitJoin(hld.fromType, ofc.val1, hld.joinL);
-				addImplicitJoin(hld.fromType, ofc.val2, hld.joinL);
-			}
-			//TODO: do like and IN filters too, and AndOr
-		}
-
-		private void addImplicitJoin(DStructType fromType, FilterVal fval, List<JoinElement> resultL) {
-			if (fval.isSymbol()) {
-				String fieldName = fval.asSymbol();
-				addFieldJoinIfNeeded(fromType, fieldName, resultL);
-			} else if (fval.isFn()) {
-				FilterFunc filterFn = fval.asFunc();
-				for(FilterVal inner: filterFn.argL) {
-					addImplicitJoin(fromType, inner, resultL); //*** recursion ***
-				}
-			}
-		}
-
-		private void addFieldJoinIfNeeded(DStructType fromType, String fieldName, List<JoinElement> resultL) {
-			TypePair pair = DRuleHelper.findMatchingStructPair(fromType, fieldName);
-			if (pair != null) {
-				JoinElement el = buildElement(fromType, fieldName, (DStructType) pair.type);
-				switch(el.relinfo.cardinality) {
-				case ONE_TO_ONE:
-				case ONE_TO_MANY:
-					if (el.relinfo.isParent) {
-						addElement(el, resultL);
-					}
-					break;
-				case MANY_TO_MANY:
-					addElement(el, resultL);
-					break;
-				}
-			}
-		}
-
-		private void addOrderBy(QueryFnSpec fnspec, List<JoinElement> resultL) {
-			DStructType structType = fnspec.structField.dtype;
-			String fieldName = fnspec.structField.fieldName;
-			TypePair pair = DRuleHelper.findMatchingPair(structType, fieldName);
-			//ignore sort by scalar fields
-			if (pair != null && pair.type instanceof DStructType) {
-				//				addElement(structType, fieldName, (DStructType) pair.type, resultL);
-				JoinElement el = buildElement(structType, fieldName, (DStructType) pair.type);
-				addElement(el, resultL);
-			}
-		}
-
-		private void addFKs(FetchSpec spec, List<JoinElement> resultL) {
-			DStructType structType = spec.structField.dtype;
-			for(TypePair pair: structType.getAllFields()) {
-				if (pair.type.isStructShape()) {
-					RelationInfo relinfo = DRuleHelper.findMatchingRuleInfo(structType, pair);
-					if (relinfo.isParent || RelationCardinality.MANY_TO_MANY.equals(relinfo.cardinality)) {
-						JoinElement el = buildElement(structType, pair.name, (DStructType) pair.type);
-						el.usedForFK = true;
-						addElement(el, resultL);
-					}
-				}
-			}
-		}
-		private void addFetch(FetchSpec spec, List<JoinElement> resultL) {
-			DStructType structType = spec.structField.dtype;
-			String fieldName = spec.structField.fieldName;
-			TypePair pair = DRuleHelper.findMatchingStructPair(structType, fieldName);
-			if (pair != null) {
-				JoinElement el = buildElement(structType, pair.name, (DStructType) pair.type);
-				el.usedForFK = false;
-				addElement(el, resultL);
-			}
-		}
-		private JoinElement buildElement(DStructType dtype, String field, DStructType fieldType) {
-			JoinElement el = new JoinElement();
-			el.structField = new RelationField(dtype, field, fieldType);
-			el.relinfo = DRuleHelper.findMatchingRuleInfo(dtype, el.createPair());
-			return el;
-		}
-		private void addElement(JoinElement el, List<JoinElement> resultL) {
-			String target = el.toString();
-			Optional<JoinElement> optExisting = resultL.stream().filter(x -> x.toString().equals(target)).findAny();
-			if (optExisting.isPresent()) {
-				if (el.usedForFK) {
-					optExisting.get().usedForFK = true; //propogate
-				}
-				if (el.usedForFetch) {
-					optExisting.get().usedForFetch = true; //propogate
-				}
-				return;
-			}
-
-			resultL.add(el);
-		}
-	}
-
-	public static class HLDField {
-		public DStructType structType;
-		public String fieldName;
-		public DType fieldType;
-		//		public boolean isAssocField;
-//		public String groupName; just use alias??
-		public String alias;
-		public String asStr;
-
-		@Override
-		public String toString() {
-			String fldType = BuiltInTypes.convertDTypeNameToDeliaName(fieldType.getName());
-			String s = String.format("%s.%s(%s,%s)", structType.getName(), fieldName, fldType, alias);
-			return s;
-		}
-	}
 
 	public static class HLDFieldBuilder {
 		private HLDAliasManager aliasMgr;
